Split plots.ts phase 1: extract workbench, sync, and list/create handlers (warren-3f46)
