+++ .canopy/agent.json
@@ -13,7 +13,7 @@
       "agent"
     ],
     "source": "builtin",
-    "provider": "anthropic",
-    "model": "claude-opus-4-7"
+    "provider": "google",
+    "model": "gemini-3.5-flash"
   }
 }
