+++ .pi/agent/auth.json
@@ -1,0 +1,1 @@
+{}
