+++ scripts/file-size-budgets.json
@@ -31,8 +31,10 @@
 		"src/server/handlers/runs.preview.test.ts": 767,
 		"src/server/handlers/plan-runs.test.ts": 1318,
 		"src/server/handlers/plot-plan-runs.test.ts": 640,
-		"src/server/handlers/plots.test.ts": 3726,
-		"src/server/handlers/plots.ts": 1652,
+		"src/server/handlers/plots.list.test.ts": 700,
+		"src/server/handlers/plots.test.ts": 2442,
+		"src/server/handlers/plots.ts": 1010,
+		"src/server/handlers/plots.workbench.test.ts": 766,
 		"src/server/handlers/runs.test.ts": 988,
 		"src/server/handlers/runs.ts": 789,
 		"src/server/integration.multi-worker.test.ts": 553,
