+++ src/server/handlers/plots.test.ts
@@ -28,23 +28,16 @@ import {
 	PlotAttachmentNotFoundError,
 	PlotIllegalStatusTransitionError,
 	PlotIntentFrozenError,
-	PlotQuestionAlreadyAnsweredError,
-	PlotQuestionNotFoundError,
 } from "../../plots/errors.ts";
 import type {
-	AnswerPlotQuestionRequest,
-	AnswerPlotQuestionResult,
 	AttachPlotRequest,
 	AttachPlotResult,
 	ChangePlotStatusRequest,
 	ChangePlotStatusResult,
-	CreatePlotRequest,
-	CreatePlotResult,
 	DetachPlotRequest,
 	DetachPlotResult,
 	EditPlotIntentRequest,
 	EditPlotIntentResult,
-	FormalizePlotResult,
 	MergePlotPrRequest,
 	MergePlotPrResult,
 	PlotAggregator,
@@ -191,24 +184,6 @@ function fakeResolver(map: Record<string, ProjectRow | null>): {
 	return { resolver, calls };
 }
 
-interface FakeCreatorCall {
-	readonly input: CreatePlotRequest;
-}
-
-function fakeCreator(result: CreatePlotResult): {
-	creator: PlotCreator;
-	calls: FakeCreatorCall[];
-} {
-	const calls: FakeCreatorCall[] = [];
-	const creator: PlotCreator = {
-		async create(input) {
-			calls.push({ input });
-			return result;
-		},
-	};
-	return { creator, calls };
-}
-
 async function seedProject(
 	repos: Repos,
 	over: Partial<ProjectRow> & { id: string },
@@ -260,585 +235,6 @@ function fakeAggregator(rows: readonly PlotSummary[]): {
 	};
 	return { agg, state };
 }
-
-function summary(over: Partial<PlotSummary>): PlotSummary {
-	return {
-		id: "pt-a",
-		name: "A",
-		status: "active",
-		intent_goal_preview: "",
-		attachments_count: 0,
-		last_event_ts: "2026-05-18T00:00:00Z",
-		last_event_actor: "user:operator",
-		project_id: "proj-a",
-		...over,
-	};
-}
-
-describe("GET /plots", () => {
-	let db: WarrenDb;
-	let repos: Repos;
-	let handle: ServeHandle | null = null;
-
-	beforeEach(async () => {
-		db = await openDatabase({ path: ":memory:" });
-		repos = createRepos(db);
-	});
-
-	afterEach(async () => {
-		if (handle) {
-			await handle.stop();
-			handle = null;
-		}
-		await db.close();
-	});
-
-	test("returns 200 { plots: [] } when no aggregator is wired (empty-deployments contract)", async () => {
-		const deps = await depsFor({ repos });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots`);
-		expect(res.status).toBe(200);
-		const body = (await res.json()) as { plots: readonly PlotSummary[] };
-		expect(body.plots).toEqual([]);
-	});
-
-	test("returns 200 { plots: [] } when the aggregator reports zero hasPlot projects", async () => {
-		const { agg } = fakeAggregator([]);
-		const deps = await depsFor({ repos, plotAggregator: agg });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots`);
-		expect(res.status).toBe(200);
-		expect(await res.text()).toBe(`{"plots":[]}`);
-	});
-
-	test("surfaces aggregator rows as-is under the `plots` key", async () => {
-		const rows = [
-			summary({ id: "pt-1", status: "active", last_event_ts: "2026-05-18T01:00:00Z" }),
-			summary({ id: "pt-2", status: "drafting", last_event_ts: "2026-05-18T00:30:00Z" }),
-		];
-		const { agg, state } = fakeAggregator(rows);
-		const deps = await depsFor({ repos, plotAggregator: agg });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots`);
-		expect(res.status).toBe(200);
-		const body = (await res.json()) as { plots: readonly PlotSummary[] };
-		expect(body.plots.map((p) => p.id)).toEqual(["pt-1", "pt-2"]);
-		expect(state.calls).toEqual([{}]);
-	});
-
-	test("passes ?status= through to the aggregator", async () => {
-		const rows = [
-			summary({ id: "pt-1", status: "active" }),
-			summary({ id: "pt-2", status: "drafting" }),
-		];
-		const { agg, state } = fakeAggregator(rows);
-		const deps = await depsFor({ repos, plotAggregator: agg });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots?status=active`);
-		expect(res.status).toBe(200);
-		const body = (await res.json()) as { plots: readonly PlotSummary[] };
-		expect(body.plots.map((p) => p.id)).toEqual(["pt-1"]);
-		expect(state.calls).toEqual([{ status: "active" }]);
-	});
-
-	test("treats empty ?status= as no filter", async () => {
-		const rows = [summary({ id: "pt-1", status: "active" })];
-		const { agg, state } = fakeAggregator(rows);
-		const deps = await depsFor({ repos, plotAggregator: agg });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots?status=`);
-		expect(res.status).toBe(200);
-		expect(state.calls).toEqual([{}]);
-	});
-
-	test("rejects unknown ?status= with 400 + validation_error", async () => {
-		const { agg, state } = fakeAggregator([]);
-		const deps = await depsFor({ repos, plotAggregator: agg });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots?status=bogus`);
-		expect(res.status).toBe(400);
-		const body = (await res.json()) as { error: { code: string; message: string } };
-		expect(body.error.code).toBe("validation_error");
-		expect(body.error.message).toContain("bogus");
-		expect(state.calls).toEqual([]);
-	});
-});
-
-/**
- * `GET /plots?filter=needs_attention` + `GET /plots/needs-attention/count`
- * — warren-d693 / pl-0344 step 9. Pins:
- *   - filter routes through `listNeedsAttention`, not `listSummaries`
- *   - status filter composes on top of needs_attention
- *   - unknown filter values are rejected with 400 validation_error
- *   - count endpoint mirrors `countNeedsAttention`
- *   - the route precedes `/plots/:id` so `needs-attention` isn't an :id
- *   - empty-deployments contract: missing aggregator ⇒ stable empty body
- */
-describe("GET /plots?filter=needs_attention", () => {
-	let db: WarrenDb;
-	let repos: Repos;
-	let handle: ServeHandle | null = null;
-
-	beforeEach(async () => {
-		db = await openDatabase({ path: ":memory:" });
-		repos = createRepos(db);
-	});
-
-	afterEach(async () => {
-		if (handle) {
-			await handle.stop();
-			handle = null;
-		}
-		await db.close();
-	});
-
-	function needsAttentionAggregator(
-		rows: ReadonlyArray<
-			PlotSummary & {
-				reasons: ReadonlyArray<"paused_run" | "merged_pr_unreviewed" | "stale_draft">;
-			}
-		>,
-	): PlotAggregator {
-		return {
-			async listSummaries() {
-				return rows;
-			},
-			async listNeedsAttention() {
-				return rows;
-			},
-			async countNeedsAttention() {
-				return rows.length;
-			},
-			invalidate() {},
-		};
-	}
-
-	test("returns aggregator listNeedsAttention rows under the `plots` key", async () => {
-		const rows = [
-			{ ...summary({ id: "pt-1", status: "active" }), reasons: ["paused_run"] as const },
-			{
-				...summary({ id: "pt-2", status: "drafting" }),
-				reasons: ["stale_draft"] as const,
-			},
-		];
-		const agg = needsAttentionAggregator(rows);
-		const deps = await depsFor({ repos, plotAggregator: agg });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots?filter=needs_attention`);
-		expect(res.status).toBe(200);
-		const body = (await res.json()) as {
-			plots: ReadonlyArray<PlotSummary & { reasons: string[] }>;
-		};
-		expect(body.plots.map((r) => r.id)).toEqual(["pt-1", "pt-2"]);
-		expect(body.plots[0]?.reasons).toEqual(["paused_run"]);
-	});
-
-	test("composes ?status= on top of ?filter=needs_attention", async () => {
-		const rows = [
-			{ ...summary({ id: "pt-1", status: "active" }), reasons: ["paused_run"] as const },
-			{
-				...summary({ id: "pt-2", status: "drafting" }),
-				reasons: ["stale_draft"] as const,
-			},
-		];
-		const agg = needsAttentionAggregator(rows);
-		const deps = await depsFor({ repos, plotAggregator: agg });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots?filter=needs_attention&status=drafting`);
-		expect(res.status).toBe(200);
-		const body = (await res.json()) as { plots: ReadonlyArray<PlotSummary> };
-		expect(body.plots.map((r) => r.id)).toEqual(["pt-2"]);
-	});
-
-	test("rejects unknown ?filter= with 400 validation_error", async () => {
-		const agg = needsAttentionAggregator([]);
-		const deps = await depsFor({ repos, plotAggregator: agg });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots?filter=bogus`);
-		expect(res.status).toBe(400);
-		const body = (await res.json()) as { error: { code: string; message: string } };
-		expect(body.error.code).toBe("validation_error");
-		expect(body.error.message).toContain("bogus");
-	});
-
-	test("empty body when aggregator is not wired", async () => {
-		const deps = await depsFor({ repos });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots?filter=needs_attention`);
-		expect(res.status).toBe(200);
-		const body = (await res.json()) as { plots: ReadonlyArray<PlotSummary> };
-		expect(body.plots).toEqual([]);
-	});
-});
-
-describe("GET /plots/needs-attention/count", () => {
-	let db: WarrenDb;
-	let repos: Repos;
-	let handle: ServeHandle | null = null;
-
-	beforeEach(async () => {
-		db = await openDatabase({ path: ":memory:" });
-		repos = createRepos(db);
-	});
-
-	afterEach(async () => {
-		if (handle) {
-			await handle.stop();
-			handle = null;
-		}
-		await db.close();
-	});
-
-	test("returns `{ count }` mirroring the aggregator", async () => {
-		const agg: PlotAggregator = {
-			async listSummaries() {
-				return [];
-			},
-			async listNeedsAttention() {
-				return [];
-			},
-			async countNeedsAttention() {
-				return 3;
-			},
-			invalidate() {},
-		};
-		const deps = await depsFor({ repos, plotAggregator: agg });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots/needs-attention/count`);
-		expect(res.status).toBe(200);
-		const body = (await res.json()) as { count: number };
-		expect(body.count).toBe(3);
-	});
-
-	test("returns `{ count: 0 }` when aggregator is not wired (empty-deployment contract)", async () => {
-		const deps = await depsFor({ repos });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots/needs-attention/count`);
-		expect(res.status).toBe(200);
-		expect(await res.text()).toBe(`{"count":0}`);
-	});
-
-	test("does NOT shadow GET /plots/:id when the param is 'needs-attention'", async () => {
-		// Because /plots/needs-attention/count is the only static path that
-		// shares the /plots/<token> prefix and the count route is registered
-		// before /plots/:id, a literal /plots/needs-attention (no /count)
-		// must still fall through to /plots/:id (which would 400 the bad id
-		// shape). This pins the ordering invariant.
-		const agg: PlotAggregator = {
-			async listSummaries() {
-				return [];
-			},
-			async listNeedsAttention() {
-				return [];
-			},
-			async countNeedsAttention() {
-				return 0;
-			},
-			invalidate() {},
-		};
-		const deps = await depsFor({ repos, plotAggregator: agg });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		// /plots/needs-attention (no /count) routes to GET /plots/:id with
-		// id='needs-attention'; the id-shape / lookup path rejects it as a
-		// 4xx (400 or 404) rather than serving an aggregator response.
-		const res = await fetch(`${tcpUrl(handle)}/plots/needs-attention`);
-		expect(res.status).toBeGreaterThanOrEqual(400);
-		expect(res.status).toBeLessThan(500);
-	});
-});
-
-describe("POST /plots", () => {
-	let db: WarrenDb;
-	let repos: Repos;
-	let handle: ServeHandle | null = null;
-
-	beforeEach(async () => {
-		db = await openDatabase({ path: ":memory:" });
-		repos = createRepos(db);
-	});
-
-	afterEach(async () => {
-		if (handle) {
-			await handle.stop();
-			handle = null;
-		}
-		await db.close();
-	});
-
-	const HAPPY_RESULT: CreatePlotResult = {
-		id: "pt-new",
-		name: "Fresh Plot",
-		status: "drafting",
-		intent_goal_preview: "ship it",
-		attachments_count: 0,
-		last_event_ts: "2026-05-18T01:23:45Z",
-		last_event_actor: "user:operator",
-	};
-
-	test("happy path: creates a Plot in a hasPlot project and returns the PlotSummary", async () => {
-		const project = await seedProject(repos, { id: "proj-plot", hasPlot: true });
-		const { creator, calls } = fakeCreator(HAPPY_RESULT);
-		const { agg, state } = fakeAggregator([]);
-		const deps = await depsFor({ repos, plotAggregator: agg, plotCreator: creator });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({
-				project_id: project.id,
-				name: "Fresh Plot",
-				intent: { goal: "ship it", non_goals: ["yak shave"] },
-				dispatcher_handle: "alice",
-			}),
-		});
-		expect(res.status).toBe(201);
-		const body = (await res.json()) as PlotSummary;
-		expect(body).toEqual({
-			id: "pt-new",
-			name: "Fresh Plot",
-			status: "drafting",
-			intent_goal_preview: "ship it",
-			attachments_count: 0,
-			last_event_ts: "2026-05-18T01:23:45Z",
-			last_event_actor: "user:operator",
-			project_id: project.id,
-		});
-
-		expect(calls).toHaveLength(1);
-		const call = calls[0];
-		if (call === undefined) throw new Error("expected one creator call");
-		expect(call.input.handle).toBe("alice");
-		expect(call.input.name).toBe("Fresh Plot");
-		expect(call.input.plotDir).toBe(`${project.localPath}/.plot`);
-		expect(call.input.intent).toEqual({ goal: "ship it", non_goals: ["yak shave"] });
-
-		// Aggregator cache was invalidated for the owning project.
-		expect(state.invalidates).toEqual([project.id]);
-	});
-
-	test("rejects when project.hasPlot=false with ProjectLacksPlotError", async () => {
-		const project = await seedProject(repos, { id: "proj-noplot", hasPlot: false });
-		const { creator, calls } = fakeCreator(HAPPY_RESULT);
-		const deps = await depsFor({ repos, plotCreator: creator });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({ project_id: project.id, name: "Won't land" }),
-		});
-		expect(res.status).toBe(400);
-		const body = (await res.json()) as { error: { code: string; message: string } };
-		expect(body.error.code).toBe("project_lacks_plot");
-		expect(body.error.message).toContain(project.id);
-		expect(calls).toEqual([]);
-	});
-
-	test("404s on unknown project_id", async () => {
-		const deps = await depsFor({ repos, plotCreator: fakeCreator(HAPPY_RESULT).creator });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({ project_id: "prj-missing", name: "x" }),
-		});
-		expect(res.status).toBe(404);
-	});
-
-	test("defaults missing name to 'Untitled Plot'", async () => {
-		const project = await seedProject(repos, { id: "proj-untitled", hasPlot: true });
-		const { creator, calls } = fakeCreator({ ...HAPPY_RESULT, name: "Untitled Plot" });
-		const deps = await depsFor({ repos, plotCreator: creator });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({ project_id: project.id }),
-		});
-		expect(res.status).toBe(201);
-		const call = calls[0];
-		if (call === undefined) throw new Error("expected one creator call");
-		expect(call.input.name).toBe("Untitled Plot");
-		expect(call.input.intent).toBeUndefined();
-	});
-
-	test("downgrades malformed dispatcher_handle to 'operator'", async () => {
-		const project = await seedProject(repos, { id: "proj-handle", hasPlot: true });
-		const { creator, calls } = fakeCreator(HAPPY_RESULT);
-		const deps = await depsFor({ repos, plotCreator: creator });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({
-				project_id: project.id,
-				name: "x",
-				dispatcher_handle: "!!not a handle!!",
-			}),
-		});
-		expect(res.status).toBe(201);
-		const call = calls[0];
-		if (call === undefined) throw new Error("expected one creator call");
-		expect(call.input.handle).toBe("operator");
-	});
-
-	test("rejects empty string name with 400", async () => {
-		const project = await seedProject(repos, { id: "proj-emptyname", hasPlot: true });
-		const { creator, calls } = fakeCreator(HAPPY_RESULT);
-		const deps = await depsFor({ repos, plotCreator: creator });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({ project_id: project.id, name: "   " }),
-		});
-		expect(res.status).toBe(400);
-		expect(calls).toEqual([]);
-	});
-
-	test("rejects unknown intent field with 400", async () => {
-		const project = await seedProject(repos, { id: "proj-badintent", hasPlot: true });
-		const { creator, calls } = fakeCreator(HAPPY_RESULT);
-		const deps = await depsFor({ repos, plotCreator: creator });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({
-				project_id: project.id,
-				name: "x",
-				intent: { goal: "ok", nongoals: ["typo"] },
-			}),
-		});
-		expect(res.status).toBe(400);
-		const body = (await res.json()) as { error: { code: string; message: string } };
-		expect(body.error.message).toContain("nongoals");
-		expect(calls).toEqual([]);
-	});
-
-	test("propagates creator errors as 500 (no fire-and-log)", async () => {
-		const project = await seedProject(repos, { id: "proj-boom", hasPlot: true });
-		const boom: PlotCreator = {
-			async create() {
-				throw new Error("disk on fire");
-			},
-		};
-		const deps = await depsFor({ repos, plotCreator: boom });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({ project_id: project.id, name: "x" }),
-		});
-		// Generic error → 500 (no typed mapping); the user sees the failure
-		// synchronously rather than the create silently succeeding-with-warning.
-		expect(res.status).toBe(500);
-	});
-});
 
 describe("GET /plots/:id", () => {
 	let db: WarrenDb;
@@ -2999,728 +2395,5 @@ describe("POST /plots/:id/attachments/:ref/merge", () => {
 		);
 		expect(res.status).toBe(404);
 		expect(calls).toEqual([]);
-	});
-});
-
-/* ----------------------------------------------------------------------- */
-/* POST /plots/:id/questions/:event_id/answer (warren-e1ac / pl-9d6a step 12) */
-/* ----------------------------------------------------------------------- */
-
-interface FakeQuestionAnswererCall {
-	readonly input: AnswerPlotQuestionRequest;
-}
-
-function fakeQuestionAnswerer(result: AnswerPlotQuestionResult): {
-	answerer: PlotQuestionAnswerer;
-	calls: FakeQuestionAnswererCall[];
-} {
-	const calls: FakeQuestionAnswererCall[] = [];
-	const answerer: PlotQuestionAnswerer = {
-		async answer(input) {
-			calls.push({ input });
-			return result;
-		},
-	};
-	return { answerer, calls };
-}
-
-function answeredEvent(over: {
-	question_id?: string;
-	text?: string;
-	at?: string;
-	actor?: string;
-}): PlotEvent {
-	return {
-		type: "question_answered",
-		actor: over.actor ?? "user:alice",
-		at: over.at ?? "2026-05-18T05:00:00Z",
-		data: {
-			question_id: over.question_id ?? "2026-05-18T04:00:00Z",
-			text: over.text ?? "ship oauth",
-		},
-	};
-}
-
-describe("POST /plots/:id/questions/:event_id/answer", () => {
-	let db: WarrenDb;
-	let repos: Repos;
-	let handle: ServeHandle | null = null;
-
-	beforeEach(async () => {
-		db = await openDatabase({ path: ":memory:" });
-		repos = createRepos(db);
-	});
-
-	afterEach(async () => {
-		if (handle) {
-			await handle.stop();
-			handle = null;
-		}
-		await db.close();
-	});
-
-	const EVENT_ID = "2026-05-18T04:00:00Z";
-
-	test("happy path: appends question_answered and returns the new event", async () => {
-		const project = await seedProject(repos, { id: "proj-q", hasPlot: true });
-		const { resolver, calls: resolverCalls } = fakeResolver({ "pt-q": project });
-		const ev = answeredEvent({ question_id: EVENT_ID, text: "ship oauth" });
-		const { answerer, calls } = fakeQuestionAnswerer({ event: ev });
-		const { agg, state } = fakeAggregator([]);
-		const deps = await depsFor({
-			repos,
-			plotAggregator: agg,
-			plotResolver: resolver,
-			plotQuestionAnswerer: answerer,
-		});
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(
-			`${tcpUrl(handle)}/plots/pt-q/questions/${encodeURIComponent(EVENT_ID)}/answer`,
-			{
-				method: "POST",
-				headers: { "content-type": "application/json" },
-				body: JSON.stringify({ answer: "ship oauth", dispatcher_handle: "alice" }),
-			},
-		);
-		expect(res.status).toBe(200);
-		const body = (await res.json()) as { event: PlotEvent };
-		expect(body.event.type).toBe("question_answered");
-		expect((body.event.data as { question_id?: string }).question_id).toBe(EVENT_ID);
-		expect((body.event.data as { text?: string }).text).toBe("ship oauth");
-
-		expect(resolverCalls).toEqual(["pt-q"]);
-		expect(calls).toHaveLength(1);
-		const call = calls[0];
-		if (call === undefined) throw new Error("expected one answerer call");
-		expect(call.input.plotId).toBe("pt-q");
-		expect(call.input.eventId).toBe(EVENT_ID);
-		expect(call.input.handle).toBe("alice");
-		expect(call.input.answer).toBe("ship oauth");
-		expect(call.input.plotDir).toBe(`${project.localPath}/.plot`);
-
-		// Aggregator cache invalidated so a follow-up list sees the
-		// refreshed last_event_ts without the 5s TTL.
-		expect(state.invalidates).toEqual([project.id]);
-	});
-
-	test("decodes URL-encoded :event_id (ISO timestamps contain `:`)", async () => {
-		const project = await seedProject(repos, { id: "proj-enc", hasPlot: true });
-		const { resolver } = fakeResolver({ "pt-q": project });
-		const { answerer, calls } = fakeQuestionAnswerer({
-			event: answeredEvent({ question_id: EVENT_ID }),
-		});
-		const deps = await depsFor({
-			repos,
-			plotResolver: resolver,
-			plotQuestionAnswerer: answerer,
-		});
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(
-			`${tcpUrl(handle)}/plots/pt-q/questions/${encodeURIComponent(EVENT_ID)}/answer`,
-			{
-				method: "POST",
-				headers: { "content-type": "application/json" },
-				body: JSON.stringify({ answer: "yes" }),
-			},
-		);
-		expect(res.status).toBe(200);
-		const call = calls[0];
-		if (call === undefined) throw new Error("expected one answerer call");
-		expect(call.input.eventId).toBe(EVENT_ID);
-	});
-
-	test("rejects missing answer with 400", async () => {
-		const project = await seedProject(repos, { id: "proj-m", hasPlot: true });
-		const { resolver } = fakeResolver({ "pt-q": project });
-		const { answerer, calls } = fakeQuestionAnswerer({
-			event: answeredEvent({ question_id: EVENT_ID }),
-		});
-		const deps = await depsFor({
-			repos,
-			plotResolver: resolver,
-			plotQuestionAnswerer: answerer,
-		});
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(
-			`${tcpUrl(handle)}/plots/pt-q/questions/${encodeURIComponent(EVENT_ID)}/answer`,
-			{
-				method: "POST",
-				headers: { "content-type": "application/json" },
-				body: JSON.stringify({}),
-			},
-		);
-		expect(res.status).toBe(400);
-		expect(calls).toEqual([]);
-	});
-
-	test("rejects empty answer with 400", async () => {
-		const project = await seedProject(repos, { id: "proj-em", hasPlot: true });
-		const { resolver } = fakeResolver({ "pt-q": project });
-		const { answerer, calls } = fakeQuestionAnswerer({
-			event: answeredEvent({ question_id: EVENT_ID }),
-		});
-		const deps = await depsFor({
-			repos,
-			plotResolver: resolver,
-			plotQuestionAnswerer: answerer,
-		});
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(
-			`${tcpUrl(handle)}/plots/pt-q/questions/${encodeURIComponent(EVENT_ID)}/answer`,
-			{
-				method: "POST",
-				headers: { "content-type": "application/json" },
-				body: JSON.stringify({ answer: "" }),
-			},
-		);
-		expect(res.status).toBe(400);
-		expect(calls).toEqual([]);
-	});
-
-	test("downgrades malformed dispatcher_handle to 'operator'", async () => {
-		const project = await seedProject(repos, { id: "proj-h", hasPlot: true });
-		const { resolver } = fakeResolver({ "pt-q": project });
-		const { answerer, calls } = fakeQuestionAnswerer({
-			event: answeredEvent({ question_id: EVENT_ID }),
-		});
-		const deps = await depsFor({
-			repos,
-			plotResolver: resolver,
-			plotQuestionAnswerer: answerer,
-		});
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(
-			`${tcpUrl(handle)}/plots/pt-q/questions/${encodeURIComponent(EVENT_ID)}/answer`,
-			{
-				method: "POST",
-				headers: { "content-type": "application/json" },
-				body: JSON.stringify({ answer: "y", dispatcher_handle: "!!nope!!" }),
-			},
-		);
-		expect(res.status).toBe(200);
-		const call = calls[0];
-		if (call === undefined) throw new Error("expected one answerer call");
-		expect(call.input.handle).toBe("operator");
-	});
-
-	test("404s when the resolver returns null (unknown plot_id)", async () => {
-		const { resolver } = fakeResolver({});
-		const { answerer, calls } = fakeQuestionAnswerer({
-			event: answeredEvent({ question_id: EVENT_ID }),
-		});
-		const deps = await depsFor({
-			repos,
-			plotResolver: resolver,
-			plotQuestionAnswerer: answerer,
-		});
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(
-			`${tcpUrl(handle)}/plots/pt-missing/questions/${encodeURIComponent(EVENT_ID)}/answer`,
-			{
-				method: "POST",
-				headers: { "content-type": "application/json" },
-				body: JSON.stringify({ answer: "y" }),
-			},
-		);
-		expect(res.status).toBe(404);
-		expect(calls).toEqual([]);
-	});
-
-	test("404s when no resolver is wired (non-Plot deployment)", async () => {
-		const { answerer, calls } = fakeQuestionAnswerer({
-			event: answeredEvent({ question_id: EVENT_ID }),
-		});
-		const deps = await depsFor({ repos, plotQuestionAnswerer: answerer });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(
-			`${tcpUrl(handle)}/plots/pt-x/questions/${encodeURIComponent(EVENT_ID)}/answer`,
-			{
-				method: "POST",
-				headers: { "content-type": "application/json" },
-				body: JSON.stringify({ answer: "y" }),
-			},
-		);
-		expect(res.status).toBe(404);
-		expect(calls).toEqual([]);
-	});
-
-	test("surfaces ProjectLacksPlotError when hasPlot flipped after resolution", async () => {
-		const project = await seedProject(repos, { id: "proj-flip", hasPlot: false });
-		const { resolver } = fakeResolver({ "pt-flip": project });
-		const { answerer, calls } = fakeQuestionAnswerer({
-			event: answeredEvent({ question_id: EVENT_ID }),
-		});
-		const deps = await depsFor({
-			repos,
-			plotResolver: resolver,
-			plotQuestionAnswerer: answerer,
-		});
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(
-			`${tcpUrl(handle)}/plots/pt-flip/questions/${encodeURIComponent(EVENT_ID)}/answer`,
-			{
-				method: "POST",
-				headers: { "content-type": "application/json" },
-				body: JSON.stringify({ answer: "y" }),
-			},
-		);
-		expect(res.status).toBe(400);
-		const body = (await res.json()) as { error: { code: string } };
-		expect(body.error.code).toBe("project_lacks_plot");
-		expect(calls).toEqual([]);
-	});
-
-	test("surfaces PlotQuestionNotFoundError from the answerer as 404 (seed-pinned: pin no-such-question)", async () => {
-		const project = await seedProject(repos, { id: "proj-nf", hasPlot: true });
-		const { resolver } = fakeResolver({ "pt-nf": project });
-		const missing: PlotQuestionAnswerer = {
-			async answer() {
-				throw new PlotQuestionNotFoundError(
-					`plot pt-nf has no question_posed event at ${EVENT_ID}`,
-				);
-			},
-		};
-		const deps = await depsFor({
-			repos,
-			plotResolver: resolver,
-			plotQuestionAnswerer: missing,
-		});
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(
-			`${tcpUrl(handle)}/plots/pt-nf/questions/${encodeURIComponent(EVENT_ID)}/answer`,
-			{
-				method: "POST",
-				headers: { "content-type": "application/json" },
-				body: JSON.stringify({ answer: "y" }),
-			},
-		);
-		expect(res.status).toBe(404);
-		const body = (await res.json()) as { error: { code: string; message: string } };
-		expect(body.error.code).toBe("plot_question_not_found");
-		expect(body.error.message).toContain(EVENT_ID);
-	});
-
-	test("surfaces PlotQuestionAlreadyAnsweredError from the answerer as 409 (seed-pinned: already-answered rejection)", async () => {
-		const project = await seedProject(repos, { id: "proj-aa", hasPlot: true });
-		const { resolver } = fakeResolver({ "pt-aa": project });
-		const already: PlotQuestionAnswerer = {
-			async answer() {
-				throw new PlotQuestionAlreadyAnsweredError(
-					`plot pt-aa question ${EVENT_ID} already has a question_answered reply`,
-				);
-			},
-		};
-		const deps = await depsFor({
-			repos,
-			plotResolver: resolver,
-			plotQuestionAnswerer: already,
-		});
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(
-			`${tcpUrl(handle)}/plots/pt-aa/questions/${encodeURIComponent(EVENT_ID)}/answer`,
-			{
-				method: "POST",
-				headers: { "content-type": "application/json" },
-				body: JSON.stringify({ answer: "y" }),
-			},
-		);
-		expect(res.status).toBe(409);
-		const body = (await res.json()) as { error: { code: string; message: string } };
-		expect(body.error.code).toBe("plot_question_already_answered");
-		expect(body.error.message).toContain(EVENT_ID);
-	});
-
-	test("propagates generic answerer errors as 500 (no fire-and-log)", async () => {
-		const project = await seedProject(repos, { id: "proj-boom", hasPlot: true });
-		const { resolver } = fakeResolver({ "pt-q": project });
-		const boom: PlotQuestionAnswerer = {
-			async answer() {
-				throw new Error("disk on fire");
-			},
-		};
-		const deps = await depsFor({
-			repos,
-			plotResolver: resolver,
-			plotQuestionAnswerer: boom,
-		});
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(
-			`${tcpUrl(handle)}/plots/pt-q/questions/${encodeURIComponent(EVENT_ID)}/answer`,
-			{
-				method: "POST",
-				headers: { "content-type": "application/json" },
-				body: JSON.stringify({ answer: "y" }),
-			},
-		);
-		expect(res.status).toBe(500);
-	});
-
-	test("agent-actor unreachability: answerer request type has no actor-kind field", () => {
-		// Compile-time pin (mx-bd4d67): the wire request to the answerer
-		// carries only a string `handle` — `question_answered` is one of
-		// the four humans-only event types per SPEC §6, and the
-		// underlying UserPlotClient hard-codes `kind: "user"`. The
-		// AgentPlotHandle.append generic excludes "question_answered"
-		// at the type level, so the agent-actor mistake is unreachable
-		// from this code path.
-		const probe: AnswerPlotQuestionRequest = {
-			plotDir: "/x/.plot",
-			plotId: "pt-x",
-			handle: "alice",
-			eventId: EVENT_ID,
-			answer: "y",
-		};
-		// @ts-expect-error — `actor` is not a field on AnswerPlotQuestionRequest
-		const _bad: AnswerPlotQuestionRequest = { ...probe, actor: { kind: "agent" } };
-		void _bad;
-		expect(Object.keys(probe).sort()).toEqual(["answer", "eventId", "handle", "plotDir", "plotId"]);
-	});
-});
-
-/* ----------------------------------------------------------------------- */
-/* POST /plots/:id/formalize (warren-d22e / pl-0344 step 8)                 */
-/* ----------------------------------------------------------------------- */
-
-describe("POST /plots/:id/formalize", () => {
-	let db: WarrenDb;
-	let repos: Repos;
-	let handle: ServeHandle | null = null;
-
-	beforeEach(async () => {
-		db = await openDatabase({ path: ":memory:" });
-		repos = createRepos(db);
-	});
-
-	afterEach(async () => {
-		if (handle) {
-			await handle.stop();
-			handle = null;
-		}
-		await db.close();
-	});
-
-	test("happy path: returns suggested intent + source message count", async () => {
-		const project = await seedProject(repos, { id: "proj-form", hasPlot: true });
-		const { resolver } = fakeResolver({ "plot-formalize01": project });
-		const formalizer: PlotFormalizer = {
-			async formalize(input) {
-				return {
-					plot_id: input.plotId,
-					suggested_intent: {
-						goal: "ship it",
-						non_goals: ["A"],
-						constraints: ["B"],
-						success_criteria: ["C"],
-					},
-					source_message_count: 3,
-				};
-			},
-		};
-		const deps = await depsFor({ repos, plotResolver: resolver, plotFormalizer: formalizer });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots/plot-formalize01/formalize`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({}),
-		});
-		expect(res.status).toBe(200);
-		const body = (await res.json()) as FormalizePlotResult;
-		expect(body.plot_id).toBe("plot-formalize01");
-		expect(body.suggested_intent.goal).toBe("ship it");
-		expect(body.suggested_intent.non_goals).toEqual(["A"]);
-		expect(body.source_message_count).toBe(3);
-	});
-
-	test("404 when plot id is unknown to the resolver", async () => {
-		const { resolver } = fakeResolver({});
-		const deps = await depsFor({ repos, plotResolver: resolver });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-		const res = await fetch(`${tcpUrl(handle)}/plots/plot-missing01/formalize`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({}),
-		});
-		expect(res.status).toBe(404);
-	});
-
-	test("400 project_lacks_plot when project hasPlot flag is false", async () => {
-		const project = await seedProject(repos, { id: "proj-noplot", hasPlot: false });
-		const { resolver } = fakeResolver({ "plot-noplot00": project });
-		const deps = await depsFor({ repos, plotResolver: resolver });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-		const res = await fetch(`${tcpUrl(handle)}/plots/plot-noplot00/formalize`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({}),
-		});
-		expect(res.status).toBe(400);
-		const body = (await res.json()) as { error: { code: string } };
-		expect(body.error.code).toBe("project_lacks_plot");
-	});
-
-	test("default formalizer reads agent_message events from runs bound to the plot", async () => {
-		// End-to-end with the default `createDefaultPlotFormalizer({repos})`:
-		// seed two interactive runs bound to the same plot, append
-		// agent_message events with marker content, and assert the
-		// suggestion folds them together.
-		const project = await seedProject(repos, { id: "proj-roundtrip", hasPlot: true });
-		const { resolver } = fakeResolver({ "plot-roundtrip0": project });
-		await repos.agents.upsert({
-			name: "brainstorm",
-			renderedJson: {
-				name: "brainstorm",
-				version: 1,
-				sections: { system: "x" },
-				resolvedFrom: [],
-				frontmatter: {},
-			},
-		});
-		const run1 = await repos.runs.create({
-			projectId: project.id,
-			agentName: "brainstorm",
-			renderedAgentJson: {},
-			prompt: "p1",
-			trigger: "brainstorm",
-			mode: "interactive",
-			plotId: "plot-roundtrip0",
-		});
-		const run2 = await repos.runs.create({
-			projectId: project.id,
-			agentName: "brainstorm",
-			renderedAgentJson: {},
-			prompt: "p2",
-			trigger: "brainstorm",
-			mode: "interactive",
-			plotId: "plot-roundtrip0",
-		});
-		await repos.events.append({
-			runId: run1.id,
-			burrowEventSeq: 1,
-			ts: "2026-05-23T00:00:00Z",
-			kind: "agent_message",
-			stream: "system",
-			payload: { content: "**goal**: roundtrip-goal\n**non_goals**:\n- A" },
-		});
-		await repos.events.append({
-			runId: run2.id,
-			burrowEventSeq: 1,
-			ts: "2026-05-23T00:01:00Z",
-			kind: "agent_message",
-			stream: "system",
-			payload: { content: "**constraints**:\n- C1\n**success_criteria**:\n- S1" },
-		});
-		// A non-agent_message event must be ignored.
-		await repos.events.append({
-			runId: run2.id,
-			burrowEventSeq: 2,
-			ts: "2026-05-23T00:01:30Z",
-			kind: "user_message",
-			stream: "system",
-			payload: { content: "**goal**: must-be-ignored" },
-		});
-
-		const deps = await depsFor({ repos, plotResolver: resolver });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-		const res = await fetch(`${tcpUrl(handle)}/plots/plot-roundtrip0/formalize`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({}),
-		});
-		expect(res.status).toBe(200);
-		const body = (await res.json()) as FormalizePlotResult;
-		expect(body.suggested_intent.goal).toBe("roundtrip-goal");
-		expect(body.suggested_intent.non_goals).toEqual(["A"]);
-		expect(body.suggested_intent.constraints).toEqual(["C1"]);
-		expect(body.suggested_intent.success_criteria).toEqual(["S1"]);
-		expect(body.source_message_count).toBe(2);
-	});
-
-	test("source_message_count is 0 when no agent_message events exist", async () => {
-		const project = await seedProject(repos, { id: "proj-fresh", hasPlot: true });
-		const { resolver } = fakeResolver({ "plot-fresh00000": project });
-		const deps = await depsFor({ repos, plotResolver: resolver });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-		const res = await fetch(`${tcpUrl(handle)}/plots/plot-fresh00000/formalize`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({}),
-		});
-		expect(res.status).toBe(200);
-		const body = (await res.json()) as FormalizePlotResult;
-		expect(body.source_message_count).toBe(0);
-		expect(body.suggested_intent).toEqual({
-			goal: "",
-			non_goals: [],
-			constraints: [],
-			success_criteria: [],
-		});
-	});
-});
-
-describe("POST /plots/:id/sync", () => {
-	let db: WarrenDb;
-	let repos: Repos;
-	let handle: ServeHandle | null = null;
-
-	beforeEach(async () => {
-		db = await openDatabase({ path: ":memory:" });
-		repos = createRepos(db);
-	});
-
-	afterEach(async () => {
-		if (handle) {
-			await handle.stop();
-			handle = null;
-		}
-		await db.close();
-	});
-
-	test("returns the synced result when successful", async () => {
-		const project = await seedProject(repos, { id: "proj-sync", hasPlot: true });
-		const { resolver } = fakeResolver({ "plot-sync-01": project });
-		const mockSyncer: import("../../plots/index.ts").PlotSyncer = {
-			async sync(_input) {
-				return {
-					kind: "synced",
-					branch: "warren/plot-sync-xyz",
-					prUrl: "https://github.com/owner/repo/pull/1",
-					merged: true,
-				};
-			},
-		};
-
-		const deps = await depsFor({ repos, plotResolver: resolver, plotSyncer: mockSyncer });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots/plot-sync-01/sync`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({}),
-		});
-
-		expect(res.status).toBe(200);
-		const body = await res.json();
-		expect(body).toEqual({
-			kind: "synced",
-			branch: "warren/plot-sync-xyz",
-			prUrl: "https://github.com/owner/repo/pull/1",
-			merged: true,
-		});
-	});
-
-	test("returns 404 for non-existent plot id", async () => {
-		const deps = await depsFor({ repos });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots/plot-missing-01/sync`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({}),
-		});
-
-		expect(res.status).toBe(404);
-	});
-
-	test("returns error when project lacks plot directory", async () => {
-		const project = await seedProject(repos, { id: "proj-noplot", hasPlot: false });
-		const { resolver } = fakeResolver({ "plot-noplot-01": project });
-		const deps = await depsFor({ repos, plotResolver: resolver });
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: NO_AUTH,
-			logger: silentLogger,
-		});
-
-		const res = await fetch(`${tcpUrl(handle)}/plots/plot-noplot-01/sync`, {
-			method: "POST",
-			headers: { "content-type": "application/json" },
-			body: JSON.stringify({}),
-		});
-
-		expect(res.status).toBe(400); // ProjectLacksPlotError -> 400
 	});
 });
