+++ src/server/handlers/plots.ts
@@ -18,31 +18,19 @@ import {
 import { NotFoundError, ValidationError } from "../../core/errors.ts";
 import { ProjectLacksPlotError } from "../../plan-runs/errors.ts";
 import {
-	createDefaultPlotFormalizer,
 	defaultPlotAttacher,
-	defaultPlotCreator,
 	defaultPlotIntentEditor,
 	defaultPlotPrMerger,
-	defaultPlotQuestionAnswerer,
 	defaultPlotReader,
 	defaultPlotRenamer,
 	defaultPlotStatusChanger,
-	defaultPlotSyncer,
-	EMPTY_PLOT_SUMMARIES,
 	type PlotEnvelope,
-	type PlotNeedsAttentionSummary,
 	type PlotSummary,
 	type PlotSummaryArtifact,
 	summarizePlot,
 } from "../../plots/index.ts";
 import { refreshProject } from "../../projects/index.ts";
-import {
-	appendUserMessage,
-	buildInteractivePrompt,
-	defaultPlotContextReader,
-	resolveDispatcherHandle,
-	spawnRun,
-} from "../../runs/index.ts";
+import { resolveDispatcherHandle } from "../../runs/index.ts";
 import { DEFAULT_AGENT_PAUSE_TIMEOUT_MS, loadWarrenConfig } from "../../warren-config/index.ts";
 import { jsonResponse } from "../response.ts";
 import type { RouteHandler, ServerDeps } from "../types.ts";
@@ -52,445 +40,8 @@ import {
 	readJsonBody,
 	readJsonBodyOrEmpty,
 	requireParam,
-	requireString,
 } from "./index.ts";
-
-/**
- * `GET /plots?status=` — list Plot summaries aggregated across every
- * `hasPlot=true` project (warren-c167 / pl-9d6a step 2).
- *
- * Single-response JSON: no NDJSON envelope here — the list is bounded
- * by the number of Plots across all hasPlot projects (Plot SPEC notes
- * this is small-N relative to runs). The 5s in-memory cache lives
- * inside the aggregator (`src/plots/aggregate.ts`); successful list
- * calls never block on a fresh per-project rebuild.
- *
- * Empty-deployments contract (pinned by `28-plot-list-and-create.ts`
- * and the unit test below): when zero projects have `hasPlot=true` we
- * return the byte-identical empty array. `EMPTY_PLOT_SUMMARIES` is the
- * canonical reference so a non-Plot deployment that walks the API
- * surface sees a stable `200 []` response without any new bytes —
- * preserving the CLAUDE.md "opt-in built-in feature" framing.
- *
- * Status filter: validated against the `@os-eco/plot-cli`
- * `PLOT_STATUSES` whitelist so a typo doesn't silently return zero
- * results. Unknown status → `ValidationError` (→ 400 / `bad_request`).
- */
-function listPlotsHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const rawStatus = ctx.url.searchParams.get("status");
-		const rawFilter = ctx.url.searchParams.get("filter");
-		let status: PlotStatus | undefined;
-		if (rawStatus !== null && rawStatus !== "") {
-			if (!(PLOT_STATUSES as readonly string[]).includes(rawStatus)) {
-				throw new ValidationError(
-					`unknown status '${rawStatus}'; expected one of ${PLOT_STATUSES.join(", ")}`,
-				);
-			}
-			status = rawStatus as PlotStatus;
-		}
-		// `?filter=needs_attention` (warren-d693 / pl-0344 step 9) routes
-		// to the needs-attention scorer and returns rows carrying a
-		// `reasons` array. `?status` + `?filter` compose: the status
-		// filter is applied client-side on top of the needs-attention
-		// set so a UI can render "drafting Plots in the Needs you view".
-		if (rawFilter !== null && rawFilter !== "") {
-			if (rawFilter !== "needs_attention") {
-				throw new ValidationError(`unknown filter '${rawFilter}'; expected one of needs_attention`);
-			}
-			if (deps.plotAggregator === undefined) {
-				return jsonResponse(200, { plots: [] as readonly PlotNeedsAttentionSummary[] });
-			}
-			const rows = await deps.plotAggregator.listNeedsAttention();
-			const filtered = status !== undefined ? rows.filter((r) => r.status === status) : rows;
-			return jsonResponse(200, { plots: filtered });
-		}
-		// Aggregator may be omitted by tests; fall back to the byte-identical
-		// empty contract so the wire shape stays the same regardless.
-		if (deps.plotAggregator === undefined) {
-			return jsonResponse(200, { plots: EMPTY_PLOT_SUMMARIES });
-		}
-		const plots = await deps.plotAggregator.listSummaries(
-			status !== undefined ? { status } : undefined,
-		);
-		return jsonResponse(200, { plots });
-	};
-}
-
-/**
- * `GET /plots/needs-attention/count` — sidebar-badge counter for the
- * deployment-wide "Needs you" view (warren-d693 / pl-0344 step 9,
- * surfaced in the UI by warren-f0e2 / step 13).
- *
- * Returns `{ count: number }`. Deployments without `plotAggregator`
- * wired (the standalone-warren empty-deployment contract) return
- * `{ count: 0 }` — byte-stable for the no-Plot path.
- */
-function needsAttentionCountHandler(deps: ServerDeps): RouteHandler {
-	return async () => {
-		if (deps.plotAggregator === undefined) {
-			return jsonResponse(200, { count: 0 });
-		}
-		const count = await deps.plotAggregator.countNeedsAttention();
-		return jsonResponse(200, { count });
-	};
-}
-
-/**
- * `POST /plots` — create a fresh Plot in the named project's `.plot/`
- * directory (warren-194e / pl-9d6a step 3).
- *
- * Handler order:
- *   (1) load project (NotFoundError → 404).
- *   (2) reject when `project.hasPlot === false` via typed
- *       `ProjectLacksPlotError` (mirrors POST /plan-runs' gate at
- *       mx-afe7e0 — same 400 envelope, stable `project_lacks_plot`
- *       code so HTTP consumers can branch).
- *   (3) resolve the dispatcher handle via `resolveDispatcherHandle`
- *       (mx-6a9788) — malformed/empty input downgrades to `operator`
- *       rather than throwing.
- *   (4) parse + validate the body's optional `intent` patch shape.
- *   (5) hand off to `deps.plotCreator` (default `defaultPlotCreator`)
- *       which opens a `UserPlotClient` against `<project>/.plot/`,
- *       calls `PlotStore.create({name})`, optionally applies the
- *       intent patch via `editIntent`, and returns the per-project
- *       `CreatePlotResult`. **Not** fire-and-log — the user is
- *       waiting on the create result, so a failure here propagates
- *       synchronously to the HTTP response (mx-92e6b3 contrasts:
- *       PlanRun's plot-append IS fire-and-log because the user is
- *       waiting on the PlanRun, not the Plot mirror).
- *   (6) invalidate the aggregator's cache entry for the project so the
- *       next `GET /plots` (or `PlotResolver.resolve` for follow-up
- *       per-Plot handlers landing later in pl-9d6a) sees the fresh
- *       Plot without waiting for the 5s TTL.
- *   (7) return 201 with the full `PlotSummary` (per-project subset +
- *       `project_id` from the resolved row).
- *
- * Body shape: `{ project_id: string, name?: string, intent?: {
- *   goal?, non_goals?, constraints?, success_criteria? },
- *   dispatcher_handle?: string }`. `name` defaults to `"Untitled
- *   Plot"` when omitted so the UI's New-Plot dialog can ship a
- *   nameless draft (the user renames later via the per-Plot intent
- *   surface); explicit empty string is rejected since
- *   `PlotStore.create` requires non-empty.
- */
-function createPlotHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const body = await readJsonBody(ctx);
-		const projectId = requireString(body, "project_id");
-		const rawName = optionalString(body, "name");
-		if (rawName !== undefined && rawName.trim().length === 0) {
-			throw new ValidationError("field 'name' must be a non-empty string when provided");
-		}
-		const name = rawName !== undefined ? rawName : "Untitled Plot";
-		const dispatcherHandle = optionalString(body, "dispatcher_handle");
-		const intent = parseIntentPatch(body.intent);
-
-		// (1) project lookup — NotFoundError → 404.
-		const project = await deps.repos.projects.require(projectId);
-
-		// (2) hasPlot gate — typed error mirrors mx-afe7e0's shape.
-		if (!project.hasPlot) {
-			throw new ProjectLacksPlotError(
-				`project ${project.id} has no .plot/ directory; cannot create a Plot`,
-				{
-					recoveryHint:
-						"run `plot init` in the project clone and refresh the project so warren picks up the .plot/ directory",
-				},
-			);
-		}
-
-		// (3) dispatcher handle resolution (mx-6a9788).
-		const handle = resolveDispatcherHandle(dispatcherHandle);
-
-		// (4/5) delegate to the creator seam.
-		const creator = deps.plotCreator ?? defaultPlotCreator;
-		const created = await creator.create({
-			plotDir: join(project.localPath, ".plot"),
-			handle,
-			name,
-			...(intent !== undefined ? { intent } : {}),
-		});
-
-		// (6) drop the aggregator's cache entry so subsequent reads see
-		// the fresh Plot without waiting for the 5s TTL.
-		deps.plotAggregator?.invalidate(project.id);
-
-		// (7) wire response — PlotSummary shape.
-		const summary: PlotSummary = {
-			id: created.id,
-			name: created.name,
-			status: created.status,
-			intent_goal_preview: created.intent_goal_preview,
-			attachments_count: created.attachments_count,
-			last_event_ts: created.last_event_ts,
-			last_event_actor: created.last_event_actor,
-			project_id: project.id,
-		};
-		return jsonResponse(201, summary);
-	};
-}
-
-/**
- * Parse + validate the optional `intent` body field on `POST /plots`.
- * `null`/`undefined` → undefined (no patch). Anything else must be a
- * JSON object; unknown fields are rejected so a typo'd `goals`/`nongoals`
- * surfaces instead of silently dropping. List fields must be arrays of
- * non-empty strings.
- */
-function parseIntentPatch(
-	raw: unknown,
-): import("../../plots/index.ts").CreatePlotIntentPatch | undefined {
-	if (raw === undefined || raw === null) return undefined;
-	if (typeof raw !== "object" || Array.isArray(raw)) {
-		throw new ValidationError("field 'intent' must be a JSON object");
-	}
-	const obj = raw as Record<string, unknown>;
-	const allowed = new Set(["goal", "non_goals", "constraints", "success_criteria"]);
-	for (const key of Object.keys(obj)) {
-		if (!allowed.has(key)) {
-			throw new ValidationError(
-				`field 'intent.${key}' is not recognized; expected one of goal, non_goals, constraints, success_criteria`,
-			);
-		}
-	}
-	const patch: {
-		goal?: string;
-		non_goals?: string[];
-		constraints?: string[];
-		success_criteria?: string[];
-	} = {};
-	if (obj.goal !== undefined) {
-		if (typeof obj.goal !== "string") {
-			throw new ValidationError("field 'intent.goal' must be a string");
-		}
-		patch.goal = obj.goal;
-	}
-	for (const key of ["non_goals", "constraints", "success_criteria"] as const) {
-		const v = obj[key];
-		if (v === undefined) continue;
-		if (!Array.isArray(v) || v.some((item) => typeof item !== "string" || item.length === 0)) {
-			throw new ValidationError(`field 'intent.${key}' must be an array of non-empty strings`);
-		}
-		patch[key] = v as string[];
-	}
-	return patch;
-}
-
-/**
- * `POST /brainstorm` — one-shot brainstorm dispatcher (warren-d22e /
- * pl-0344 step 8).
- *
- * Atomically creates a draft Plot (status defaults to `drafting`, empty
- * intent, name defaults to `"Untitled brainstorm"`) and dispatches the
- * first interactive turn against the built-in `brainstorm` agent bound
- * to that Plot. The two operations are exposed as separate primitives
- * (`POST /plots`, `POST /runs` with `mode='interactive'`) but the
- * fresh-start UX is a single click — wrapping them here saves the UI
- * an awkward two-call dance plus a partial-failure rollback.
- *
- * Body shape: `{ project_id: string, prompt: string, name?: string,
- *   agent?: string, dispatcher_handle?: string, providerOverride?,
- *   modelOverride?, ref? }`. `agent` overrides the default
- *   `"brainstorm"` so canopy-library variants can swap in by name; the
- *   override flows through warren's normal agent-resolution path.
- *
- * Handler order mirrors `POST /plots` then `POST /runs` (interactive):
- *   (1) project lookup + `hasPlot` gate (`ProjectLacksPlotError`).
- *   (2) dispatcher-handle resolution (mx-6a9788).
- *   (3) create the draft Plot via `deps.plotCreator`. Failure surfaces
- *       synchronously — the user is waiting on the result; we don't
- *       want a half-state where the run lands without a Plot.
- *   (4) build the first-turn prompt via `buildInteractivePrompt` (the
- *       same envelope `spawnInteractiveTurn` emits on turn N, so the
- *       agent sees a uniform shape across turn 1 and turn N).
- *   (5) `spawnRun({mode:'interactive', plotId: created.id, agentName})`.
- *   (6) `appendUserMessage` on the new run so the events stream
- *       reflects the conversation from turn 0.
- *   (7) start the bridge so events flow into warren immediately.
- *   (8) return 201 with `{plot, run, burrow}`.
- *
- * Aggregator cache is invalidated after the create so subsequent
- * `GET /plots` lists see the fresh Plot without waiting for the 5s
- * TTL.
- */
-function createBrainstormHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const body = await readJsonBody(ctx);
-		const projectId = requireString(body, "project_id");
-		const prompt = requireString(body, "prompt");
-		const rawName = optionalString(body, "name");
-		if (rawName !== undefined && rawName.trim().length === 0) {
-			throw new ValidationError("field 'name' must be a non-empty string when provided");
-		}
-		const name = rawName !== undefined ? rawName : "Untitled brainstorm";
-		const dispatcherHandle = optionalString(body, "dispatcher_handle");
-		const providerOverride = optionalString(body, "providerOverride");
-		const modelOverride = optionalString(body, "modelOverride");
-		const ref = optionalString(body, "ref");
-		const agentName = optionalString(body, "agent") ?? "brainstorm";
-
-		// (1) project + hasPlot gate.
-		const project = await deps.repos.projects.require(projectId);
-		if (!project.hasPlot) {
-			throw new ProjectLacksPlotError(
-				`project ${project.id} has no .plot/ directory; cannot start a brainstorm`,
-				{
-					recoveryHint:
-						"run `plot init` in the project clone and refresh the project so warren picks up the .plot/ directory",
-				},
-			);
-		}
-
-		// (2) dispatcher-handle resolution.
-		const handle = resolveDispatcherHandle(dispatcherHandle);
-
-		// (3) create the draft Plot.
-		const creator = deps.plotCreator ?? defaultPlotCreator;
-		const created = await creator.create({
-			plotDir: join(project.localPath, ".plot"),
-			handle,
-			name,
-		});
-		deps.plotAggregator?.invalidate(project.id);
-
-		// (4) build first-turn prompt envelope. Best-effort context load —
-		// a freshly-created Plot has no events yet, but we still go through
-		// the same builder so the agent sees a uniform shape.
-		let context = null;
-		try {
-			context = await defaultPlotContextReader.read({
-				plotDir: join(project.localPath, ".plot"),
-				plotId: created.id,
-				historyTail: 0,
-				handle,
-			});
-		} catch {
-			// Best-effort; matches the createRunHandler interactive-first-turn
-			// posture above.
-		}
-		const dispatchedPrompt = buildInteractivePrompt(context, prompt);
-
-		// (5) dispatch the first interactive turn.
-		const spawned = await spawnRun({
-			repos: deps.repos,
-			burrowClientPool: deps.burrowClientPool,
-			agentName,
-			projectId,
-			prompt: dispatchedPrompt,
-			mode: "interactive",
-			plotId: created.id,
-			dispatcherHandle: handle,
-			trigger: "brainstorm",
-			projectsConfig: deps.projectsConfig,
-			projectSpawn: deps.spawn ?? defaultSpawn,
-			...(ref !== undefined ? { ref } : {}),
-			...(providerOverride !== undefined ? { providerOverride } : {}),
-			...(modelOverride !== undefined ? { modelOverride } : {}),
-			...(deps.now !== undefined ? { now: deps.now } : {}),
-			...(deps.warrenConfigs !== undefined ? { warrenConfigs: deps.warrenConfigs } : {}),
-			...(deps.runBranchPrefixDefault !== undefined
-				? { runBranchPrefixDefault: deps.runBranchPrefixDefault }
-				: {}),
-			...(deps.seedsCli !== undefined ? { seedsCli: deps.seedsCli } : {}),
-		});
-
-		// (6) append the raw user_message event on turn 0.
-		await appendUserMessage({
-			repos: deps.repos,
-			runId: spawned.run.id,
-			message: prompt,
-			handle,
-			...(deps.now !== undefined ? { now: deps.now() } : {}),
-		});
-
-		// (7) start the bridge so events flow into warren immediately.
-		deps.bridges.start(spawned.run.id, spawned.burrowRun.id, spawned.burrow.id);
-
-		// (8) wire response.
-		const summary: PlotSummary = {
-			id: created.id,
-			name: created.name,
-			status: created.status,
-			intent_goal_preview: created.intent_goal_preview,
-			attachments_count: created.attachments_count,
-			last_event_ts: created.last_event_ts,
-			last_event_actor: created.last_event_actor,
-			project_id: project.id,
-		};
-		return jsonResponse(201, {
-			plot: summary,
-			run: spawned.run,
-			burrow: {
-				id: spawned.burrow.id,
-				workspacePath: spawned.burrow.workspacePath,
-			},
-		});
-	};
-}
-
-/**
- * `POST /plots/:id/formalize` — brainstorm-summarize endpoint
- * (warren-d22e / pl-0344 step 8).
- *
- * Returns a **suggested** Plot intent extracted from the agent's
- * messages in the brainstorm conversation. Non-mutating: the Plot is
- * untouched, no Plot event is emitted, no run is spawned. The caller
- * (UI) renders the suggestion as a form the user edits, then applies
- * via the existing `POST /plots/:id/intent` route and transitions to
- * `ready` via `POST /plots/:id/status`.
- *
- * Handler order:
- *   (1) resolve owning project via `deps.plotResolver`. `null` → typed
- *       404 so the empty-deployments contract stays stable.
- *   (2) defensive `hasPlot` re-check — the resolver only walks
- *       `hasPlot=true` rows, but the flag can flip out from under us.
- *   (3) hand off to `deps.plotFormalizer` (default
- *       `createDefaultPlotFormalizer({repos})`) which reads every
- *       `agent_message` event on interactive runs bound to the Plot,
- *       parses field markers, and returns the accumulated suggestion.
- *
- * Response shape: `{plot_id, suggested_intent: {goal, non_goals,
- *   constraints, success_criteria}, source_message_count}`. An
- *   all-empty suggestion + `source_message_count: 0` is the legitimate
- *   "no agent_message events yet" response — the UI shows "start
- *   chatting first" instead of an empty form.
- */
-function formalizePlotHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const plotId = requireParam(ctx, "id");
-
-		// (1) resolve owning project.
-		const project =
-			deps.plotResolver !== undefined ? await deps.plotResolver.resolve(plotId) : null;
-		if (project === null) {
-			throw new NotFoundError(`plot not found: ${plotId}`, {
-				recoveryHint:
-					"check the plot id; only Plots in projects with hasPlot=true are visible to warren",
-			});
-		}
-
-		// (2) defensive hasPlot re-check.
-		if (!project.hasPlot) {
-			throw new ProjectLacksPlotError(
-				`project ${project.id} no longer has a .plot/ directory; cannot formalize plot ${plotId}`,
-				{
-					recoveryHint:
-						"refresh the project so warren picks up the current .plot/ state, or recreate .plot/ via `plot init`",
-				},
-			);
-		}
-
-		// (3) delegate to the formalizer seam.
-		const formalizer = deps.plotFormalizer ?? createDefaultPlotFormalizer({ repos: deps.repos });
-		const result = await formalizer.formalize({ plotId });
-
-		// Trigger background sync
-		triggerBackgroundSync(deps, project, plotId);
-
-		return jsonResponse(200, result);
-	};
-}
+import { triggerBackgroundSync } from "./plots/sync.ts";
 
 /**
  * Resolve `paused_runs[]` for a Plot envelope (warren-4ea4 /
@@ -1440,213 +991,13 @@ function mergePlotPrAttachmentHandler(deps: ServerDeps): RouteHandler {
 	};
 }
 
-/**
- * `POST /plots/:id/questions/:event_id/answer` — answer a `question_posed`
- * event with a `question_answered` event (warren-e1ac / pl-9d6a step 12).
- *
- * The wire `:event_id` is the `at` ISO timestamp of the targeted
- * `question_posed` event. PlotEvent has no synthetic id; `at` is the
- * stable identifier the UI feeds back after reading the event log.
- *
- * Handler order:
- *   (1) decode `:event_id` (the router runs `decodeURIComponent`) and
- *       reject empty values at the edge.
- *   (2) parse + validate the body's `answer` field (non-empty string).
- *       `dispatcher_handle` resolution via `resolveDispatcherHandle`
- *       (mx-6a9788) — malformed/empty input downgrades to `operator`.
- *   (3) resolve the owning project via `deps.plotResolver`; `null` or
- *       unwired resolver → 404 (empty-deployments contract).
- *   (4) defensive `hasPlot` re-check — surfaces as
- *       `ProjectLacksPlotError` (400 / `project_lacks_plot`) for parity
- *       with the create / intent / status / attach paths.
- *   (5) hand off to `deps.plotQuestionAnswerer` (default
- *       `defaultPlotQuestionAnswerer`). The answerer re-validates the
- *       handler-edge concurrency invariant against the fresh on-disk
- *       event log (the targeted `question_posed` still exists AND has
- *       no subsequent `question_answered` referencing it) and appends
- *       the `question_answered` event via the typed `UserPlotClient`.
- *       Failure propagates synchronously — NOT fire-and-log; the user
- *       is waiting on the answer submit.
- *   (6) invalidate the aggregator cache entry for the project so a
- *       follow-up `GET /plots` sees the refreshed `last_event_ts` /
- *       `last_event_actor` without the 5s TTL wait.
- *   (7) return 200 with `{ event: PlotEvent }` — the freshly appended
- *       `question_answered` event for the UI to splice into the
- *       optimistic activity feed; the next `GET /plots/:id` reconciles
- *       the rest of the envelope.
- *
- * ACL note: this handler uses `UserPlotClient` exclusively (via the
- * answerer seam). `question_answered` is one of the four humans-only
- * event types per SPEC §6 — `AgentPlotHandle.append` excludes it at
- * the type level (mx-bd4d67), so the agent-actor mistake is
- * unreachable from this code path at compile time.
- */
-function answerPlotQuestionHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const plotId = requireParam(ctx, "id");
-		const eventId = requireParam(ctx, "event_id");
-		if (eventId.length === 0) {
-			throw new ValidationError("path param ':event_id' must be a non-empty string");
-		}
-
-		const body = await readJsonBody(ctx);
-		const dispatcherHandle = optionalString(body, "dispatcher_handle");
-
-		// (2) parse + validate `answer`.
-		const rawAnswer = body.answer;
-		if (typeof rawAnswer !== "string" || rawAnswer.length === 0) {
-			throw new ValidationError("field 'answer' must be a non-empty string");
-		}
-
-		// handle resolution.
-		const handle = resolveDispatcherHandle(dispatcherHandle);
-
-		// (3) resolve owning project.
-		const project =
-			deps.plotResolver !== undefined ? await deps.plotResolver.resolve(plotId) : null;
-		if (project === null) {
-			throw new NotFoundError(`plot not found: ${plotId}`, {
-				recoveryHint:
-					"check the plot id; only Plots in projects with hasPlot=true are visible to warren",
-			});
-		}
-
-		// (4) defensive hasPlot re-check.
-		if (!project.hasPlot) {
-			throw new ProjectLacksPlotError(
-				`project ${project.id} no longer has a .plot/ directory; cannot answer question on plot ${plotId}`,
-				{
-					recoveryHint:
-						"refresh the project so warren picks up the current .plot/ state, or recreate .plot/ via `plot init`",
-				},
-			);
-		}
-
-		// (5) delegate to the answerer seam. The answerer re-reads the
-		// event log and re-runs `assertQuestionAnswerable` against the
-		// fresh on-disk state before calling `append` — the lib
-		// guarantees neither half of the invariant, so warren owns it.
-		const answerer = deps.plotQuestionAnswerer ?? defaultPlotQuestionAnswerer;
-		const result = await answerer.answer({
-			plotDir: join(project.localPath, ".plot"),
-			plotId,
-			handle,
-			eventId,
-			answer: rawAnswer,
-		});
-
-		// (6) drop the aggregator cache so the next list sees the
-		// refreshed last_event_ts / last_event_actor without the 5s TTL.
-		deps.plotAggregator?.invalidate(project.id);
-
-		// (7) wire response — just the freshly appended event for
-		// optimistic UI splice.
-		return jsonResponse(200, { event: result.event });
-	};
-}
-
-function triggerBackgroundSync(
-	deps: ServerDeps,
-	project: { id: string; localPath: string; gitUrl: string; defaultBranch: string },
-	plotId: string,
-): void {
-	const syncer = deps.plotSyncer ?? defaultPlotSyncer;
-	const token = deps.autoOpenPr?.token ?? "";
-	void (async () => {
-		const config =
-			deps.warrenConfigs !== undefined
-				? await deps.warrenConfigs.get(project.id, project.localPath)
-				: await loadWarrenConfig({ projectPath: project.localPath });
-		return syncer.sync({
-			projectPath: project.localPath,
-			gitUrl: project.gitUrl,
-			defaultBranch: project.defaultBranch,
-			token,
-			handle: "warren",
-			plotSyncConfig: config.defaults?.plotSync,
-			spawn: deps.spawn ?? defaultSpawn,
-			gitBinary: deps.projectsConfig.gitBinary,
-		});
-	})()
-		.then((result) => {
-			deps.logger.info({ projectId: project.id, plotId, result }, "background plot sync complete");
-		})
-		.catch((err) => {
-			deps.logger.error(
-				{
-					projectId: project.id,
-					plotId,
-					error: err instanceof Error ? err.message : String(err),
-				},
-				"background plot sync failed",
-			);
-		});
-}
-
-function syncPlotHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const plotId = requireParam(ctx, "id");
-
-		const project =
-			deps.plotResolver !== undefined ? await deps.plotResolver.resolve(plotId) : null;
-		if (project === null) {
-			throw new NotFoundError(`plot not found: ${plotId}`, {
-				recoveryHint:
-					"check the plot id; only Plots in projects with hasPlot=true are visible to warren",
-			});
-		}
-
-		if (!project.hasPlot) {
-			throw new ProjectLacksPlotError(
-				`project ${project.id} no longer has a .plot/ directory; cannot sync plot ${plotId}`,
-				{
-					recoveryHint:
-						"refresh the project so warren picks up the current .plot/ state, or recreate .plot/ via `plot init`",
-				},
-			);
-		}
-
-		const syncer = deps.plotSyncer ?? defaultPlotSyncer;
-		const token = deps.autoOpenPr?.token ?? "";
-		let plotSyncConfig: import("../../warren-config/index.ts").PlotSyncConfig | undefined;
-		try {
-			const config =
-				deps.warrenConfigs !== undefined
-					? await deps.warrenConfigs.get(project.id, project.localPath)
-					: await loadWarrenConfig({ projectPath: project.localPath });
-			plotSyncConfig = config.defaults?.plotSync;
-		} catch {
-			// Config unavailable (e.g. clone missing) — proceed with defaults.
-		}
-		const result = await syncer.sync({
-			projectPath: project.localPath,
-			gitUrl: project.gitUrl,
-			defaultBranch: project.defaultBranch,
-			token,
-			handle: "warren",
-			plotSyncConfig,
-			spawn: deps.spawn ?? defaultSpawn,
-			gitBinary: deps.projectsConfig.gitBinary,
-		});
-
-		return jsonResponse(200, result);
-	};
-}
-
 export {
-	answerPlotQuestionHandler,
 	attachPlotHandler,
 	changePlotStatusHandler,
-	createBrainstormHandler,
-	createPlotHandler,
 	detachPlotHandler,
 	editPlotIntentHandler,
-	formalizePlotHandler,
 	getPlotHandler,
 	getPlotSummaryHandler,
-	listPlotsHandler,
 	mergePlotPrAttachmentHandler,
-	needsAttentionCountHandler,
 	renamePlotHandler,
-	syncPlotHandler,
 };
