+++ src/server/handlers/plots/index.ts
@@ -1,0 +1,28 @@
+/**
+ * Composite export for the plots handlers domain (warren-3f46 / pl-3255 step 1).
+ */
+
+export {
+	attachPlotHandler,
+	changePlotStatusHandler,
+	detachPlotHandler,
+	editPlotIntentHandler,
+	getPlotHandler,
+	getPlotSummaryHandler,
+	mergePlotPrAttachmentHandler,
+	renamePlotHandler,
+} from "../plots.ts";
+export {
+	createPlotHandler,
+	listPlotsHandler,
+	needsAttentionCountHandler,
+} from "./list.ts";
+export {
+	syncPlotHandler,
+	triggerBackgroundSync,
+} from "./sync.ts";
+export {
+	answerPlotQuestionHandler,
+	createBrainstormHandler,
+	formalizePlotHandler,
+} from "./workbench.ts";
